rename generated-cwl to cwl_v1_2
